Update faces.h
Updated frame filetype to byte (uint8_t), as matrix is not larger than 255 in any dimension.
changed KILL to 255 to remain within byte.

Removed animation struct, replaced with 2 arrays _anim and _animdelay.
To reference the location in the master array "arrayface" and the correlating time the frame is up, respectively.
also included currentframe, which will be shared between all animations. reset to 0 at the end of each animation action.